Combine bench_sign and bench_verify into single bench
