+++ .gitignore
@@ -1,9 +1,7 @@
-bench_inv
+bench
 bench_ecdh
 bench_ecmult
 bench_schnorrsig
-bench_sign
-bench_verify
 bench_recover
 bench_internal
 tests
